Changes to clustering and plotting of heatmap for paper Constantin et al, 2021
Remove effectors that are not present in any isolate to prevent clustering error, when FoEC.py is called with -e flag.
Indicate whether an effector is located on the core genome of Fo47 (gene description has to contain Fo47core).
Indicate whether an effector is located on the accessory genome of Fo47 (gene description has to contain Fo47accesssory).
Indicate whether a gene product is an enzyme (gene name ends on 'ase').